add `AbstractCollection`
